+++ core/src/spotify_id.rs
@@ -21,6 +21,7 @@ pub enum SpotifyItemType {
     Playlist,
     Show,
     Track,
+    Local,
     Unknown,
 }
 
@@ -33,6 +34,7 @@ impl From<&str> for SpotifyItemType {
             "playlist" => Self::Playlist,
             "show" => Self::Show,
             "track" => Self::Track,
+            "local" => Self::Local,
             _ => Self::Unknown,
         }
     }
@@ -47,6 +49,7 @@ impl From<SpotifyItemType> for &str {
             SpotifyItemType::Playlist => "playlist",
             SpotifyItemType::Show => "show",
             SpotifyItemType::Track => "track",
+            SpotifyItemType::Local => "local",
             _ => "unknown",
         }
     }
@@ -167,24 +170,45 @@ impl SpotifyId {
     ///
     /// [Spotify URI]: https://developer.spotify.com/documentation/web-api/#spotify-uris-and-ids
     pub fn from_uri(src: &str) -> SpotifyIdResult {
-        let mut uri_parts: Vec<&str> = src.split(':').collect();
+        // Basic: `spotify:{type}:{id}`
+        // Named: `spotify:user:{user}:{type}:{id}`
+        // Local: `spotify:local:{artist}:{album_title}:{track_title}:{duration_in_seconds}`
+        let mut parts = src.split(':');
 
-        // At minimum, should be `spotify:{type}:{id}`
-        if uri_parts.len() < 3 {
-            return Err(SpotifyIdError::InvalidFormat.into());
-        }
+        let scheme = parts.next().ok_or(SpotifyIdError::InvalidFormat)?;
 
-        if uri_parts[0] != "spotify" {
+        let item_type = {
+            let next = parts.next().ok_or(SpotifyIdError::InvalidFormat)?;
+            if next == "user" {
+                let _username = parts.next().ok_or(SpotifyIdError::InvalidFormat)?;
+                parts.next().ok_or(SpotifyIdError::InvalidFormat)?
+            } else {
+                next
+            }
+        };
+
+        let id = parts.next().ok_or(SpotifyIdError::InvalidFormat)?;
+
+        if scheme != "spotify" {
             return Err(SpotifyIdError::InvalidRoot.into());
         }
 
-        let id = uri_parts.pop().unwrap_or_default();
+        let item_type = item_type.into();
+
+        // Local files have a variable-length ID: https://developer.spotify.com/documentation/general/guides/local-files-spotify-playlists/
+        // TODO: find a way to add this local file ID to SpotifyId.
+        // One possible solution would be to copy the contents of `id` to a new String field in SpotifyId,
+        // but then we would need to remove the derived Copy trait, which would be a breaking change.
+        if item_type == SpotifyItemType::Local {
+            return Ok(Self { item_type, id: 0 });
+        }
+
         if id.len() != Self::SIZE_BASE62 {
             return Err(SpotifyIdError::InvalidId.into());
         }
 
         Ok(Self {
-            item_type: uri_parts.pop().unwrap_or_default().into(),
+            item_type,
             ..Self::from_base62(id)?
         })
     }
@@ -321,6 +345,7 @@ impl NamedSpotifyId {
         let mut dst = String::with_capacity(37 + self.username.len() + item_type.len());
         dst.push_str("spotify:user:");
         dst.push_str(&self.username);
+        dst.push(':');
         dst.push_str(item_type);
         dst.push(':');
         let base_62 = self.to_base62()?;
@@ -399,12 +424,12 @@ impl TryFrom<&Vec<u8>> for SpotifyId {
 impl TryFrom<&protocol::spirc::TrackRef> for SpotifyId {
     type Error = crate::Error;
     fn try_from(track: &protocol::spirc::TrackRef) -> Result<Self, Self::Error> {
-        match SpotifyId::from_raw(track.get_gid()) {
+        match SpotifyId::from_raw(track.gid()) {
             Ok(mut id) => {
                 id.item_type = SpotifyItemType::Track;
                 Ok(id)
             }
-            Err(_) => SpotifyId::from_uri(track.get_uri()),
+            Err(_) => SpotifyId::from_uri(track.uri()),
         }
     }
 }
@@ -414,7 +439,7 @@ impl TryFrom<&protocol::metadata::Album> for SpotifyId {
     fn try_from(album: &protocol::metadata::Album) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Album,
-            ..Self::from_raw(album.get_gid())?
+            ..Self::from_raw(album.gid())?
         })
     }
 }
@@ -424,7 +449,7 @@ impl TryFrom<&protocol::metadata::Artist> for SpotifyId {
     fn try_from(artist: &protocol::metadata::Artist) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Artist,
-            ..Self::from_raw(artist.get_gid())?
+            ..Self::from_raw(artist.gid())?
         })
     }
 }
@@ -434,7 +459,7 @@ impl TryFrom<&protocol::metadata::Episode> for SpotifyId {
     fn try_from(episode: &protocol::metadata::Episode) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Episode,
-            ..Self::from_raw(episode.get_gid())?
+            ..Self::from_raw(episode.gid())?
         })
     }
 }
@@ -444,7 +469,7 @@ impl TryFrom<&protocol::metadata::Track> for SpotifyId {
     fn try_from(track: &protocol::metadata::Track) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Track,
-            ..Self::from_raw(track.get_gid())?
+            ..Self::from_raw(track.gid())?
         })
     }
 }
@@ -454,7 +479,7 @@ impl TryFrom<&protocol::metadata::Show> for SpotifyId {
     fn try_from(show: &protocol::metadata::Show) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Show,
-            ..Self::from_raw(show.get_gid())?
+            ..Self::from_raw(show.gid())?
         })
     }
 }
@@ -464,7 +489,7 @@ impl TryFrom<&protocol::metadata::ArtistWithRole> for SpotifyId {
     fn try_from(artist: &protocol::metadata::ArtistWithRole) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Artist,
-            ..Self::from_raw(artist.get_artist_gid())?
+            ..Self::from_raw(artist.artist_gid())?
         })
     }
 }
@@ -474,7 +499,7 @@ impl TryFrom<&protocol::playlist4_external::Item> for SpotifyId {
     fn try_from(item: &protocol::playlist4_external::Item) -> Result<Self, Self::Error> {
         Ok(Self {
             item_type: SpotifyItemType::Track,
-            ..Self::from_uri(item.get_uri())?
+            ..Self::from_uri(item.uri())?
         })
     }
 }
@@ -484,7 +509,7 @@ impl TryFrom<&protocol::playlist4_external::Item> for SpotifyId {
 impl TryFrom<&protocol::playlist4_external::MetaItem> for SpotifyId {
     type Error = crate::Error;
     fn try_from(item: &protocol::playlist4_external::MetaItem) -> Result<Self, Self::Error> {
-        Self::try_from(item.get_revision())
+        Self::try_from(item.revision())
     }
 }
 
@@ -494,7 +519,7 @@ impl TryFrom<&protocol::playlist4_external::SelectedListContent> for SpotifyId {
     fn try_from(
         playlist: &protocol::playlist4_external::SelectedListContent,
     ) -> Result<Self, Self::Error> {
-        Self::try_from(playlist.get_revision())
+        Self::try_from(playlist.revision())
     }
 }
 
@@ -506,7 +531,7 @@ impl TryFrom<&protocol::playlist_annotate3::TranscodedPicture> for SpotifyId {
     fn try_from(
         picture: &protocol::playlist_annotate3::TranscodedPicture,
     ) -> Result<Self, Self::Error> {
-        Self::from_base62(picture.get_uri())
+        Self::from_base62(picture.uri())
     }
 }
 
@@ -534,7 +559,7 @@ mod tests {
         raw: &'static [u8],
     }
 
-    static CONV_VALID: [ConversionCase; 4] = [
+    static CONV_VALID: [ConversionCase; 5] = [
         ConversionCase {
             id: 238762092608182713602505436543891614649,
             kind: SpotifyItemType::Track,
@@ -574,6 +599,14 @@ mod tests {
             raw: &[
                 154, 27, 28, 251, 198, 242, 68, 86, 154, 224, 53, 108, 119, 187, 233, 216,
             ],
+        },
+        ConversionCase {
+            id: 0,
+            kind: SpotifyItemType::Local,
+            uri: "spotify:local:0000000000000000000000",
+            base16: "00000000000000000000000000000000",
+            base62: "0000000000000000000000",
+            raw: &[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
         },
     ];
 
@@ -674,6 +707,25 @@ mod tests {
         for c in &CONV_INVALID {
             assert!(SpotifyId::from_uri(c.uri).is_err());
         }
+    }
+
+    #[test]
+    fn from_local_uri() {
+        let actual = SpotifyId::from_uri("spotify:local:xyz:123").unwrap();
+
+        assert_eq!(actual.id, 0);
+        assert_eq!(actual.item_type, SpotifyItemType::Local);
+    }
+
+    #[test]
+    fn from_named_uri() {
+        let actual =
+            NamedSpotifyId::from_uri("spotify:user:spotify:playlist:37i9dQZF1DWSw8liJZcPOI")
+                .unwrap();
+
+        assert_eq!(actual.id, 136159921382084734723401526672209703396);
+        assert_eq!(actual.item_type, SpotifyItemType::Playlist);
+        assert_eq!(actual.username, "spotify");
     }
 
     #[test]