+++ playback/src/decoder/passthrough_decoder.rs
@@ -14,7 +14,7 @@ use crate::{
     MS_PER_PAGE, PAGES_PER_MS,
 };
 
-fn get_header<T>(code: u8, rdr: &mut PacketReader<T>) -> DecoderResult<Box<[u8]>>
+fn get_header<T>(code: u8, rdr: &mut PacketReader<T>) -> DecoderResult<Vec<u8>>
 where
     T: Read + Seek,
 {
@@ -29,19 +29,19 @@ where
         return Err(DecoderError::PassthroughDecoder("Invalid Data".into()));
     }
 
-    Ok(pck.data.into_boxed_slice())
+    Ok(pck.data)
 }
 
 pub struct PassthroughDecoder<R: Read + Seek> {
     rdr: PacketReader<R>,
-    wtr: PacketWriter<Vec<u8>>,
+    wtr: PacketWriter<'static, Vec<u8>>,
     eos: bool,
     bos: bool,
     ofsgp_page: u64,
     stream_serial: u32,
-    ident: Box<[u8]>,
-    comment: Box<[u8]>,
-    setup: Box<[u8]>,
+    ident: Vec<u8>,
+    comment: Vec<u8>,
+    setup: Vec<u8>,
 }
 
 impl<R: Read + Seek> PassthroughDecoder<R> {
@@ -49,8 +49,7 @@ impl<R: Read + Seek> PassthroughDecoder<R> {
     pub fn new(rdr: R, format: AudioFileFormat) -> DecoderResult<Self> {
         if !AudioFiles::is_ogg_vorbis(format) {
             return Err(DecoderError::PassthroughDecoder(format!(
-                "Passthrough decoder is not implemented for format {:?}",
-                format
+                "Passthrough decoder is not implemented for format {format:?}"
             )));
         }
 
@@ -60,7 +59,7 @@ impl<R: Read + Seek> PassthroughDecoder<R> {
             .map_err(|e| DecoderError::PassthroughDecoder(e.to_string()))?;
         let stream_serial = since_epoch.as_millis() as u32;
 
-        info!("Starting passthrough track with serial {}", stream_serial);
+        info!("Starting passthrough track with serial {stream_serial}");
 
         // search for ident, comment, setup
         let ident = get_header(1, &mut rdr)?;
@@ -99,7 +98,7 @@ impl<R: Read + Seek> AudioDecoder for PassthroughDecoder<R> {
                     let absgp_page = pck.absgp_page() - self.ofsgp_page;
                     self.wtr
                         .write_packet(
-                            pck.data.into_boxed_slice(),
+                            pck.data,
                             self.stream_serial,
                             PacketWriteEndInfo::EndStream,
                             absgp_page,
@@ -197,7 +196,7 @@ impl<R: Read + Seek> AudioDecoder for PassthroughDecoder<R> {
 
             self.wtr
                 .write_packet(
-                    pck.data.into_boxed_slice(),
+                    pck.data,
                     self.stream_serial,
                     inf,
                     pckgp_page - self.ofsgp_page,