+++ playback/src/player.rs
@@ -1311,7 +1311,7 @@ impl Future for PlayerInternal {
                                                 self.send_event(PlayerEvent::PositionCorrection {
                                                     play_request_id,
                                                     track_id,
-                                                    position_ms: new_stream_position_ms as u32,
+                                                    position_ms: new_stream_position_ms,
                                                 });
                                             }
                                         }
@@ -1545,9 +1545,11 @@ impl PlayerInternal {
                         // dynamic method, there may still be peaks that we want to shave off.
 
                         // No matter the case we apply volume attenuation last if there is any.
-                        if !self.config.normalisation && volume < 1.0 {
-                            for sample in data.iter_mut() {
-                                *sample *= volume;
+                        if !self.config.normalisation {
+                            if volume < 1.0 {
+                                for sample in data.iter_mut() {
+                                    *sample *= volume;
+                                }
                             }
                         } else if self.config.normalisation_method == NormalisationMethod::Basic
                             && (normalisation_factor < 1.0 || volume < 1.0)