+++ CHANGELOG.md
@@ -38,7 +38,7 @@ https://github.com/librespot-org/librespot
 - [all] Improve lock contention and performance (breaking)
 - [audio] Files are now downloaded over the HTTPS CDN (breaking)
 - [audio] Improve file opening and seeking performance (breaking)
-- [chore] MSRV is now 1.61 (breaking)
+- [core] MSRV is now 1.65 (breaking)
 - [connect] `DeviceType` moved out of `connect` into `core` (breaking)
 - [connect] Update and expose all `spirc` context fields (breaking)
 - [connect] Add `Clone, Defaut` traits to `spirc` contexts
@@ -56,6 +56,7 @@ https://github.com/librespot-org/librespot
   now follows the setting in the Connect client that controls it. (breaking)
 - [metadata] Most metadata is now retrieved with the `spclient` (breaking)
 - [metadata] Playlists are moved to the `playlist4_external` protobuf (breaking)
+- [metadata] Handle playlists that are sent with microsecond-based timestamps
 - [playback] The audio decoder has been switched from `lewton` to `Symphonia`.
   This improves the Vorbis sound quality, adds support for MP3 as well as for
   FLAC in the future. (breaking)
@@ -68,6 +69,7 @@ https://github.com/librespot-org/librespot
 
 - [all] Check that array indexes are within bounds (panic safety)
 - [all] Wrap errors in librespot `Error` type (breaking)
+- [connect] Add option on which zeroconf will bind. Defaults to all interfaces. Ignored by DNS-SD.
 - [connect] Add session events
 - [connect] Add `repeat`, `set_position_ms` and `set_volume` to `spirc.rs`
 - [contrib] Add `event_handler_example.py`
@@ -83,6 +85,8 @@ https://github.com/librespot-org/librespot
   It supports a lot of functionality, including audio previews and image
   downloads even if librespot doesn't use that for playback itself.
 - [core] Support downloading of lyrics
+- [core] Support parsing `SpotifyId` for local files
+- [core] Support parsing `SpotifyId` for named playlists
 - [main] Add all player events to `player_event_handler.rs`
 - [main] Add an event worker thread that runs async to the main thread(s) but
   sync to itself to prevent potential data races for event consumers
@@ -91,6 +95,8 @@ https://github.com/librespot-org/librespot
   disabled such content. Applications that use librespot as a library without
   Connect should use the 'filter-explicit-content' user attribute in the session.
 - [playback] Add metadata support via a `TrackChanged` event
+- [connect] Add `activate` and `load` functions to `Spirc`, allowing control over local connect sessions
+- [metadata] Add `Lyrics`
 
 ### Fixed
 
@@ -106,6 +112,8 @@ https://github.com/librespot-org/librespot
   from the beginning
 - [playback] Handle disappearing and invalid devices better
 - [playback] Handle seek, pause, and play commands while loading
+- [playback] Handle disabled normalisation correctly when using fixed volume
+- [metadata] Fix missing colon when converting named spotify IDs to URIs
 
 ## [0.4.2] - 2022-07-29